esp32cam/requirements.txt
Flask
cloudinary
firebase-admin
Pillow
requests